requirements.txt
# Google APIs
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# WordPress / WooCommerce (REST API)
requests>=2.28.0
python-dotenv>=1.0.0
